HFile pretty printer fix; and some misc cleanup of int to long
Summary:
(i) If -m (meta only) is the only option requested, the HFile pretty printer
was still scanning the whole file.

(ii) Changed a bunch of ints to longs in HFile for future proofing. The HFile
FixedFileTrailer still has the number of keys as an "int". This is harmless. The
code actually does support larger than 2G keys, it is just that the stats
reporting can be a little off. Deferring the persistent change for future.

Test Plan:
Will run unit tests.

Ran HFile pretty printer on a large test file, and confirmed it wasn't going
into the long scan loop.

DiffCamp Revision: 175196
Reviewed By: nspiegelberg
CC: nspiegelberg
Revert Plan:
OK

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89@1181423 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/hadoop/hbase/regionserver/Store.java
@@ -507,7 +507,7 @@ public class Store implements HeapSize {
   /*
    * @return Writer for a new StoreFile in the tmp dir.
    */
-  private StoreFile.Writer createWriterInTmp(int maxKeyCount)
+  private StoreFile.Writer createWriterInTmp(long maxKeyCount)
   throws IOException {
     return StoreFile.createWriter(this.fs, region.getTmpDir(), this.blocksize,
         this.compression, this.comparator, this.conf,
@@ -844,13 +844,13 @@ public class Store implements HeapSize {
                                final boolean majorCompaction, final long maxId)
       throws IOException {
     // calculate maximum key count after compaction (for blooms)
-    int maxKeyCount = 0;
+    long maxKeyCount = 0;
     for (StoreFile file : filesToCompact) {
       StoreFile.Reader r = file.getReader();
       if (r != null) {
         // NOTE: getFilterEntries could cause under-sized blooms if the user
         //       switches bloom type (e.g. from ROW to ROWCOL)
-        int keyCount = (r.getBloomFilterType() == family.getBloomFilterType())
+        long keyCount = (r.getBloomFilterType() == family.getBloomFilterType())
             ? r.getFilterEntries() : r.getEntries();
         maxKeyCount += keyCount;
         LOG.info("Compacting: " + file + "; keyCount = " + keyCount + "; Bloom Type = " + r.getBloomFilterType().toString());
